// src/app/form2/form-data.interface.ts
export interface Formulario {
    etapa: string;
    nombrepersonal: string;
    seccionsindical: number;
    tipoCentro: string;
    centrotrabajo: string;
    evaluador: string;
    rfc: string;
    inicioperiodo: string;
    finperiodo: string;
    periodoevaluado: string;
    municipio: string;
    funcion: string;
    clavecentro: string;
    telefonocentro: string;
    dias: number;
    observaciones: string;
    cursos: number;
    tt: number;
    observasiones: string;
  
    puntajeComp: number[];
}